@@ -39,8 +39,6 @@
     "dayjs": "^1.8.20",
     "detect-it": "^3.0.5",
     "dompurify": "^2.2.2",
-    "dotenv": "^8.2.0",
-    "dotenv-webpack": "^1.7.0",
     "express": "^4.17.1",
     "google-map-react": "^1.1.6",
     "graphql": "^14.5.8",
@@ -52,7 +50,7 @@
     "lodash.debounce": "^4.0.8",
     "lodash.groupby": "^4.6.0",
     "lodash.sortby": "^4.7.0",
-    "next": "^10.2.3",
+    "next": "^11.1.4",
     "next-cookies": "^2.0.1",
     "next-i18next": "^4.0.0",
     "next-pwa": "^3.1.1",
@@ -75,34 +73,19 @@
     "redux": "^5.0.1",
     "redux-thunk": "^2.3.0",
     "reset-css": "^5.0.1",
-    "sass": "^1.90.0",
     "shopify-buy": "^2.9.1"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-optional-chaining": "^7.21.0",
-    "@babel/preset-env": "^7.28.3",
-    "@babel/preset-react": "^7.27.1",
-    "@babel/preset-typescript": "^7.27.1",
     "@testing-library/jest-dom": "^5.11.0",
     "@testing-library/react": "^10.4.3",
     "@types/jest": "^26.0.3",
     "@types/node": "^14.14.45",
     "@types/react": "17.0.2",
     "@types/react-dom": "17.0.2",
-    "@typescript-eslint/eslint-plugin": "^2.5.0",
     "@typescript-eslint/parser": "^2.5.0",
-    "babel-jest": "^30.0.5",
-    "babel-plugin-module-resolver": "^4.0.0",
     "coveralls": "^3.1.0",
-    "css-minimizer-webpack-plugin": "^7.0.2",
-    "eslint": "^6.6.0",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^6.5.0",
-    "eslint-plugin-import": "^2.18.2",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-prettier": "^3.1.1",
-    "eslint-plugin-react": "^7.16.0",
-    "fork-ts-checker-webpack-plugin": "^3.1.1",
+    "eslint": "^8.57.1",
+    "eslint-config-next": "^15.5.0",
     "husky": "^4.2.5",
     "image-trace-loader": "^1.0.2",
     "jest": "^26.1.0",
@@ -112,10 +95,10 @@
     "prettier": "^1.18.2",
     "redux-devtools-extension": "^2.13.8",
     "remote-redux-devtools": "^0.5.16",
+    "sass": "^1.90.0",
     "stylelint": "^13.6.0",
     "stylelint-config-standard": "^20.0.0",
     "tailwindcss": "^1.1.3",
-    "terser-webpack-plugin": "^5.3.14",
     "typescript": "^4.9.5",
     "url-loader": "^3.0.0"
   },
